Sor_Agent/sor_agent/requirements.txt
openai
python-dotenv
transformers
torch
scikit-learn
anthropic
